Moved previous `parser` package back to `languages/parser`
```
$ go test  ./...
ok  	github.com/chris-ramon/graphql-go	0.177s
?   	github.com/chris-ramon/graphql-go/examples/http	[no test files]
?   	github.com/chris-ramon/graphql-go/gqlerrors	[no test files]
?   	github.com/chris-ramon/graphql-go/language/ast	[no test files]
?   	github.com/chris-ramon/graphql-go/language/kinds	[no test files]
ok  	github.com/chris-ramon/graphql-go/language/lexer	0.017s
?   	github.com/chris-ramon/graphql-go/language/location	[no test files]
ok  	github.com/chris-ramon/graphql-go/language/parser	0.028s
ok  	github.com/chris-ramon/graphql-go/language/printer	0.038s
?   	github.com/chris-ramon/graphql-go/language/source	[no test files]
ok  	github.com/chris-ramon/graphql-go/language/visitor	0.019s
```
