+++ language/parser/schema_parser_test.go
@@ -1,4 +1,4 @@
-package graphql
+package parser
 
 import (
 	"reflect"
