+++ language/visitor/visitor_test.go
@@ -9,14 +9,15 @@ import (
 
 	"github.com/chris-ramon/graphql-go"
 	"github.com/chris-ramon/graphql-go/language/ast"
+	"github.com/chris-ramon/graphql-go/language/parser"
 	"github.com/chris-ramon/graphql-go/language/visitor"
 	"github.com/kr/pretty"
 )
 
 func parse(t *testing.T, query string) *ast.Document {
-	astDoc, err := graphql.Parse(graphql.ParseParams{
+	astDoc, err := parser.Parse(parser.ParseParams{
 		Source: query,
-		Options: graphql.ParseOptions{
+		Options: parser.ParseOptions{
 			NoLocation: true,
 		},
 	})
